#91 edge case, also, squelch an FZRetail warning from #215
git-svn-id: file:///media/data/aa/dev/Allegiance-SVN/branch/FAZR6@962 8c327b6c-a73e-8745-8738-111f1e176ae1

+++ src/engine/dinput.cpp
@@ -313,7 +313,7 @@ public:
         {
             ::RegQueryValueEx(hKey, "MouseSensitivity", NULL, &dwType, (unsigned char*)&szValue, &cbValue);
 
-            m_sensitivity = (strlen(szValue) >= 1 && strcmp(szValue,"0") == -1) ?  atof(szValue) : 0.83f;
+            m_sensitivity = (float)(strlen(szValue) >= 1 && strcmp(szValue,"0") == -1) ?  atof(szValue) : 0.83f;
 
             ::RegQueryValueEx(hKey, "MouseAcceleration", NULL, &dwType, (unsigned char*)&dwValue, &cwValue);
             ::RegCloseKey(hKey);
